requirements.txt
ultralytics
Flask
python-telegram-bot
python-dotenv
supervision
gTTS